refactor(folders): better structure
