+++ src/components/MultiTileSet/MultiAnswersPage.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import CardTile from './CardTile';
-import AnswerTile from './AnswerTile';
+import CardTile from 'components/TileSet/CardTile';
+import AnswerTile from 'components/TileSet/AnswerTile';
 import 'stylesheets/TileSet.css';
 
 const MultiAnswersPage = (props) => {
-  const { question, correct, answers } = props.multipleAnswers[props.current];
+  const { question, correct, answers } = props.card;
 
   const handleClick = (e) => {
     if (!props.chosenAnswer) {
