+++ src/components/ResultPage/Score.js
