+++ src/components/TileSet/LoadingAnimation.js
