+++ src/containers/MultiTileSetCont.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import MultiTileSet from 'components/MultiTileSet';
+import MultiTileSet from 'components/MultiTileSet/MultiTileSet';
 import { updateCurrentQuestion, updateChosenAnswer } from 'actions/actions';
 import { withFirebase } from 'components/Firebase';
 import { getCardsThunk } from 'thunks/thunks';
